pyproject.toml
[tool.poetry]
name = "chenv"
version = "0.1.0"
description = "modern local environment management"
authors = ["Jonathan Shemer <i@jonathanshemer.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/jonathan-shemer/chenv"
repository = "https://github.com/jonathan-shemer/chenv"
documentation = "https://chenv.readthedocs.io"
keywords = ["chenv", "dotenv", "enviroment"]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
click = "^7.1.2"
marshmallow = "^3.10.0"
desert = "^2020.11.18"
python-dotenv = "^0.15.0"
httpx = "^0.16.1"
toolz = "^0.11.1"
questionary = "^1.9.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
coverage = {extras = ["toml"], version = "^5.3.1"}
pytest-cov = "^2.8.1"
pytest-mock = "^3.4.0"
flake8 = "^3.8.4"
black = "^20.8b1"
flake8-black = "^0.2.1"
flake8-import-order = "^0.18.1"
flake8-bugbear = "^20.11.1"
flake8-bandit = "^2.1.2"
safety = "^1.10.0"
mypy = "^0.790"
typeguard = "^2.10.0"
flake8-docstrings = "^1.5.0"
darglint = "^1.5.8"
xdoctest = "^0.15.0"
sphinx = "^3.4.1"
sphinx-autodoc-typehints = "^1.11.1"
codecov = "^2.1.11"
pylint = "^2.6.0"
flake8-annotations = "^2.4.1"
requests = "^2.25.1"
pytype = "^2020.12.23"

[tool.poetry.scripts]
chenv = "chenv.main:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["chenv"]

[tool.coverage.report]
show_missing = true
# fail_under = 100

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
